implement SimplifiedModelControllers library
