+++ src/CMakeLists.txt
@@ -6,3 +6,4 @@ add_subdirectory(libYARP_helper)
 add_subdirectory(libiDynTree_helper)
 add_subdirectory(libTime_profiler)
 add_subdirectory(libStd_helper)
+add_subdirectory(libSimplified_model_controllers)
